requirements.txt
supabase==2.20.0
pillow
numpy
tensorflow
